requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
pillow==11.2.1
pycairo==1.28.0
pyscard==2.2.2
pystray==0.19.5
python-xlib==0.33
requests==2.32.3
six==1.17.0
urllib3==2.4.0
